Add apache logs to project in apachelogs dir. When apache is set in siteconfig config.json always use proxy template despite what nginx.conf says

+++ docker/nginx/run.sh
@@ -12,6 +12,7 @@ handleMedia () {
     fi
     if [ ! -d "/data/shared/media/$1" ]; then
         mkdir -p "/data/shared/media/$1"
+        chmod +w /data/shared/media/$1
     fi
     # move existing pub/media information to /data/shared/media/sitename
     # and remove origin pub/media folder
@@ -108,7 +109,7 @@ for d in `find -L /data/shared/sites -mindepth 1 -maxdepth 1 -type d`; do
         INCLUDE_PARAMS="include \/data\/shared\/sites\/$SITEBASENAME\/.siteconfig\/params.conf;"
     fi
 
-    if [ -f "/data/shared/sites/$SITEBASENAME/.siteconfig/nginx.conf" ]; then
+    if [ -f "/data/shared/sites/$SITEBASENAME/.siteconfig/nginx.conf" && $USE_WEBSERVER == "nginx" ]; then
         # Custom nginx found, don't care what the config is, uses this one.
         cp /data/shared/sites/$SITEBASENAME/.siteconfig/nginx.conf /etc/nginx/sites-enabled/$SITEBASENAME.conf
     # https://stackoverflow.com/questions/43158140/way-to-create-multiline-comments-in-bash
